# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Audiobook Organizer API",
    "version": "0.219.1",
    "description": "REST API for the audiobook organizer."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Relative path (production and same-origin clients)"
    },
    {
      "url": "http://localhost:8484/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ao_session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/maintenance/jobs/{job_id}": {
      "parameters": [
        {
          "name": "job_id",
          "in": "path",
          "required": true,
          "description": "Registered maintenance job identifier. GET /maintenance/jobs returns the live catalogue this enum is a snapshot of.",
          "schema": {
            "type": "string",
            "enum": [
              "backfill-book-files",
              "backfill-file-hashes",
              "backfill-itunes-positions",
              "backfill-metadata-source-hash",
              "backfill-sync-ids",
              "bulk-deluge-import",
              "bulk-fetch-metadata",
              "cleanup-backups",
              "cleanup-empty-folders",
              "cleanup-organize-mess",
              "cleanup-series",
              "dedup-books",
              "enrich-book-files",
              "fix-author-narrator-swap",
              "fix-book-file-paths",
              "fix-file-modes",
              "fix-read-by-narrator",
              "fix-version-groups",
              "generate-itl-tests",
              "merge-chapter-groups",
              "normalize-primary-flags",
              "purge-unknown-author-duplicates",
              "recompute-book-aggregates",
              "recompute-itunes-paths",
              "refetch-missing-authors",
              "relink-missing-to-itunes",
              "relink-report",
              "repair-missing-files",
              "retention-and-hygiene",
              "revert-metadata-fetch",
              "scan-chapter-groups",
              "scan-composer-tags",
              "scan-duplicate-files",
              "scan-duration-mismatch",
              "scan-metadata-hash-dups",
              "sweep-pebble-metrics-ttl"
            ]
          }
        }
      ],
      "post": {
        "operationId": "runMaintenanceJob",
        "summary": "Run Maintenance Job",
        "description": "Enqueues the named maintenance job as an async operation. Replaces the 15 individual POST /maintenance/{job-name} paths this spec previously (and incorrectly) advertised as separate endpoints: there is exactly one dispatch route, keyed by job_id, for every registered job. Omitting dry_run applies the job's own advertised default (see the default_params.dry_run field GET /maintenance/jobs returns for each job) rather than defaulting to false, so a client that trusts the catalogue and posts no body gets a preview, not a mutation, for jobs that default to dry_run true.",
        "tags": [
          "maintenance"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dry_run": {
                    "type": "boolean",
                    "description": "Preview instead of apply. Omitted falls back to the job's advertised default, not to false."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job enqueued, or merged into an already-active run of the same job with the same dry_run",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "operation_id": {
                      "type": "string",
                      "description": "ID of the enqueued (or merged-into) operation."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request: malformed JSON body"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: caller lacks the job's required permission"
          },
          "404": {
            "description": "Unknown job_id"
          },
          "409": {
            "description": "Conflict enqueueing the operation"
          }
        }
      }
    },
...
  }
}